Move Parallels test into parallels_test.go
